refactor: vote_crud.ts を vote_grade / vote_statistics に分割し型・依存を整理
- vote_crud.ts を vote_grade.ts（書き込み）と vote_statistics.ts（読み込み）に分割
- vote_crud.test.ts も対応する2ファイルに分割
- upsertVoteGradeTables を小関数に分解、MIN_VOTES_FOR_STATISTICS を定数化
- grade を string から TaskGrade に型強化 (vote_actions.ts)
- VoteStatisticsMap / VoteStatisticsEntry の named type を追加
- vote_statistics.ts の全関数に明示的な戻り値型を追加
- validateAdminAccess を _utils/auth.ts から再利用 (vote_management)
- getMedianVote をサービス層 (getVoteStatsByTaskId) 経由に変更
