+++ src/features/votes/services/vote_statistics.test.ts
@@ -3,36 +3,29 @@ import { describe, test, expect, vi, beforeEach } from 'vitest';
 import { TaskGrade } from '@prisma/client';
 
 import {
-  getVoteGrade,
   getVoteGradeStatistics,
   getAllTasksWithVoteInfo,
   getVoteCountersByTaskId,
   getVoteStatsByTaskId,
   getAllVoteStatisticsAsArray,
   getAllVoteCounters,
-  upsertVoteGradeTables,
-} from './vote_crud';
+} from './vote_statistics';
 
 vi.mock('$lib/server/database', () => ({
   default: {
     voteGrade: {
       findUnique: vi.fn(),
-      upsert: vi.fn(),
     },
     votedGradeCounter: {
       findMany: vi.fn(),
-      update: vi.fn(),
-      upsert: vi.fn(),
     },
     votedGradeStatistics: {
       findMany: vi.fn(),
       findFirst: vi.fn(),
-      upsert: vi.fn(),
     },
     task: {
       findMany: vi.fn(),
     },
-    $transaction: vi.fn(),
   },
 }));
 
@@ -46,7 +39,6 @@ beforeEach(() => {
 // Type aliases
 // ---------------------------------------------------------------------------
 
-type PrismaVoteGrade = Awaited<ReturnType<typeof prisma.voteGrade.findUnique>>;
 type PrismaVotedGradeStatistics = Awaited<
   ReturnType<typeof prisma.votedGradeStatistics.findMany>
 >[number];
@@ -57,10 +49,6 @@ type PrismaVotedGradeCounter = Awaited<
 // ---------------------------------------------------------------------------
 // Mock helpers
 // ---------------------------------------------------------------------------
-
-function mockVoteGradeFindUnique(value: PrismaVoteGrade) {
-  vi.mocked(prisma.voteGrade.findUnique).mockResolvedValue(value);
-}
 
 function mockVotedGradeStatisticsFindMany(value: PrismaVotedGradeStatistics[]) {
   vi.mocked(prisma.votedGradeStatistics.findMany).mockResolvedValue(
@@ -78,19 +66,6 @@ function mockTaskFindMany(value: object[]) {
   vi.mocked(prisma.task.findMany).mockResolvedValue(
     value as unknown as Awaited<ReturnType<typeof prisma.task.findMany>>,
   );
-}
-
-/** Creates a mock transaction client and wires prisma.$transaction to execute the callback with it. */
-function setupTransaction() {
-  const mockTx = {
-    voteGrade: { findUnique: vi.fn(), upsert: vi.fn() },
-    votedGradeCounter: { update: vi.fn(), upsert: vi.fn(), findMany: vi.fn() },
-    votedGradeStatistics: { upsert: vi.fn() },
-  };
-  vi.mocked(prisma.$transaction).mockImplementation(async (callback: unknown) =>
-    (callback as (tx: typeof mockTx) => Promise<unknown>)(mockTx),
-  );
-  return mockTx;
 }
 
 function makeStatisticsRecord(
@@ -124,43 +99,6 @@ function makeCounterRecord(
 // ---------------------------------------------------------------------------
 // Tests
 // ---------------------------------------------------------------------------
-
-describe('getVoteGrade', () => {
-  test('returns voted: false and grade: null when no vote record exists', async () => {
-    mockVoteGradeFindUnique(null);
-
-    const result = await getVoteGrade('user-1', 'abc001_a');
-
-    expect(result).toEqual({ voted: false, grade: null });
-  });
-
-  test('returns voted: true and the stored grade when a vote record exists', async () => {
-    mockVoteGradeFindUnique({
-      id: 'vote-1',
-      userId: 'user-1',
-      taskId: 'abc001_a',
-      grade: TaskGrade.Q5,
-      createdAt: new Date(),
-      updatedAt: new Date(),
-    });
-
-    const result = await getVoteGrade('user-1', 'abc001_a');
-
-    expect(result).toEqual({ voted: true, grade: TaskGrade.Q5 });
-  });
-
-  test('queries with the correct userId and taskId', async () => {
-    mockVoteGradeFindUnique(null);
-
-    await getVoteGrade('user-42', 'abc123_d');
-
-    expect(prisma.voteGrade.findUnique).toHaveBeenCalledWith(
-      expect.objectContaining({
-        where: { userId_taskId: { userId: 'user-42', taskId: 'abc123_d' } },
-      }),
-    );
-  });
-});
 
 describe('getVoteGradeStatistics', () => {
   test('returns an empty Map when no statistics exist', async () => {
@@ -307,92 +245,5 @@ describe('getAllVoteCounters', () => {
 
     expect(result).toHaveLength(2);
     expect(prisma.votedGradeCounter.findMany).toHaveBeenCalledWith();
-  });
-});
-
-describe('upsertVoteGradeTables', () => {
-  test('exits early without any writes when vote grade is unchanged (idempotency)', async () => {
-    const tx = setupTransaction();
-    tx.voteGrade.findUnique.mockResolvedValue({ grade: TaskGrade.Q5 });
-
-    await upsertVoteGradeTables('user-1', 'abc001_a', TaskGrade.Q5);
-
-    expect(tx.votedGradeCounter.update).not.toHaveBeenCalled();
-    expect(tx.voteGrade.upsert).not.toHaveBeenCalled();
-    expect(tx.votedGradeCounter.upsert).not.toHaveBeenCalled();
-  });
-
-  test('decrements the old grade counter when the user changes their vote', async () => {
-    const tx = setupTransaction();
-    tx.voteGrade.findUnique.mockResolvedValue({ grade: TaskGrade.Q4 });
-    tx.voteGrade.upsert.mockResolvedValue({});
-    tx.votedGradeCounter.upsert.mockResolvedValue({});
-    tx.votedGradeCounter.findMany.mockResolvedValue([
-      { grade: TaskGrade.Q5, count: 1 },
-      { grade: TaskGrade.Q4, count: 0 },
-    ]);
-
-    await upsertVoteGradeTables('user-1', 'abc001_a', TaskGrade.Q5);
-
-    expect(tx.votedGradeCounter.update).toHaveBeenCalledWith(
-      expect.objectContaining({
-        where: { taskId_grade: { taskId: 'abc001_a', grade: TaskGrade.Q4 } },
-        data: { count: { decrement: 1 } },
-      }),
-    );
-  });
-
-  test('upserts VoteGrade and increments counter for the new grade', async () => {
-    const tx = setupTransaction();
-    tx.voteGrade.findUnique.mockResolvedValue(null);
-    tx.voteGrade.upsert.mockResolvedValue({});
-    tx.votedGradeCounter.upsert.mockResolvedValue({});
-    tx.votedGradeCounter.findMany.mockResolvedValue([{ grade: TaskGrade.Q5, count: 1 }]);
-
-    await upsertVoteGradeTables('user-1', 'abc001_a', TaskGrade.Q5);
-
-    expect(tx.voteGrade.upsert).toHaveBeenCalledWith(
-      expect.objectContaining({
-        where: { userId_taskId: { userId: 'user-1', taskId: 'abc001_a' } },
-        update: { grade: TaskGrade.Q5 },
-      }),
-    );
-    expect(tx.votedGradeCounter.upsert).toHaveBeenCalledWith(
-      expect.objectContaining({
-        where: { taskId_grade: { taskId: 'abc001_a', grade: TaskGrade.Q5 } },
-        update: { count: { increment: 1 } },
-      }),
-    );
-  });
-
-  test('upserts VotedGradeStatistics when total votes reaches 3', async () => {
-    const tx = setupTransaction();
-    tx.voteGrade.findUnique.mockResolvedValue(null);
-    tx.voteGrade.upsert.mockResolvedValue({});
-    tx.votedGradeCounter.upsert.mockResolvedValue({});
-    // 3 votes all on Q5 → median = Q5
-    tx.votedGradeCounter.findMany.mockResolvedValue([{ grade: TaskGrade.Q5, count: 3 }]);
-    tx.votedGradeStatistics.upsert.mockResolvedValue({});
-
-    await upsertVoteGradeTables('user-1', 'abc001_a', TaskGrade.Q5);
-
-    expect(tx.votedGradeStatistics.upsert).toHaveBeenCalledWith(
-      expect.objectContaining({
-        where: { taskId: 'abc001_a' },
-        update: { grade: TaskGrade.Q5 },
-      }),
-    );
-  });
-
-  test('does not upsert VotedGradeStatistics when total votes is below 3', async () => {
-    const tx = setupTransaction();
-    tx.voteGrade.findUnique.mockResolvedValue(null);
-    tx.voteGrade.upsert.mockResolvedValue({});
-    tx.votedGradeCounter.upsert.mockResolvedValue({});
-    tx.votedGradeCounter.findMany.mockResolvedValue([{ grade: TaskGrade.Q5, count: 2 }]);
-
-    await upsertVoteGradeTables('user-1', 'abc001_a', TaskGrade.Q5);
-
-    expect(tx.votedGradeStatistics.upsert).not.toHaveBeenCalled();
   });
 });
