+++ src/features/votes/types/vote_result.ts
@@ -5,3 +5,9 @@ export type VoteGradeResult = {
   voted: boolean;
   grade: TaskGrade | null;
 };
+
+/** A single entry in the vote statistics map, keyed by taskId. */
+export type VoteStatisticsEntry = { grade: string };
+
+/** Map from taskId to its vote statistics entry. */
+export type VoteStatisticsMap = Map<string, VoteStatisticsEntry>;
